Move validator module one level up
